+++ CHANGES.md
@@ -21,7 +21,8 @@
 - Alt power on eye animations
 - More expressive power on anims
 - Face image overlays (Code used from WireOS)
-- Toggle between Anki lights and wireOS lights! (Run `touch /data/data/wirelights` over ssh)
+- Toggle between Anki lights and wireOS lights! (Run `touch /data/data/rebuild/wirelights` over ssh)
+- Custom backpack lights! (Following [this](https://github.com/os-vector/wire-os-victor/pull/30) pr)
 - Screen doesn't flash as much when booting up (ThommoMC's fix)
 - Add back the "Pew" sound when petting ([Reverts this Anki Commit](https://github.com/kercre123/victor/commit/48344a779ad6be70e398b96f3c79db069263e8a1))
 - Add old 1.5-era voice command response timing ([Reverts this anki commit](https://github.com/kercre123/victor/commit/6c3df37c6f3d929cc1562be0572185f10575858d))